2023.02.07
* general to include project name to avoid name clashes
* fish up ad-hoc scripts to create orgs and projects

+++ k8s/bin/create_key.bash
@@ -1,19 +1,21 @@
 #!/bin/bash
 
-adminUser="$(     kubectl get secret admin-user-credentials -o json | jq .data.Username |         sed -e's/"//g'| base64 --decode )"
-publicApiKey="$(  kubectl get secret mongodb-opsmanager-admin-key -o json | jq .data.publicKey  | sed -e's/"//g'| base64 --decode )"
-privateApiKey="$( kubectl get secret mongodb-opsmanager-admin-key -o json | jq .data.privateKey | sed -e's/"//g'| base64 --decode )"
-
-initconf=$( sed -e '/adminUser/d' -e '/private.*Key/d' -e '/public.*Key/d'  init.conf )
-printf "%s\n" "$initconf" > init.conf
-
-#echo  adminUser="${adminUser}" | tee -a init.conf
-echo  publicApiKey="${publicApiKey}"   | tee -a init.conf
-echo  privateApiKey="${privateApiKey}" | tee -a init.conf
-
 source init.conf
-file=/tmp/key.json
-rm "${file}" > /dev/null 2>&1
+source custom.conf
+
+adminUser="$(     kubectl get secret admin-user-credentials -o json | jq .data.Username |         sed -e's/"//g'| base64 --decode )"
+publicApiKey="$(  kubectl get secret ${namespace}-opsmanager-admin-key -o json | jq .data.publicKey  | sed -e's/"//g'| base64 --decode )"
+privateApiKey="$( kubectl get secret ${namespace}-opsmanager-admin-key -o json | jq .data.privateKey | sed -e's/"//g'| base64 --decode )"
+
+    conf=$( sed -e '/adminUser/d' -e '/privateApiKey/d' -e '/publicApiKey/d'  custom.conf )
+    printf "%s\n" "$conf" > custom.conf
+    echo  publicApiKey="${publicApiKey}"   | tee custom.conf
+    echo  privateApiKey="${privateApiKey}" | tee -a custom.conf
+
+if [[ $publicKey == "" ]]
+then
+    file=/tmp/key.json
+    rm "${file}" > /dev/null 2>&1
 curl --insecure --user "${publicApiKey}:${privateApiKey}" --digest \
   --header "Accept: application/json" \
   --header "Content-Type: application/json" \
@@ -26,9 +28,14 @@ curl --insecure --user "${publicApiKey}:${privateApiKey}" --digest \
 
 if [[ -e "${file}" ]]
 then
-    echo  publicKey="$(  cat ${file} |jq .publicKey  )" | tee -a init.conf
-    echo  privateKey="$( cat ${file} |jq .privateKey )" | tee -a init.conf
+    echo  publicKey="$(  cat ${file} |jq .publicKey  )" | tee -a custom.conf
+    echo  privateKey="$( cat ${file} |jq .privateKey )" | tee -a custom.conf
 else
     printf "%s\n" "Error did not create key"
     exit 1
+fi
+
+else
+    echo  publicKey="$publicKey"
+    echo  privateKey="$privateKey"
 fi
